Refactored execute.py to script.py
To signify that it is the main file of the script module.
