+++ mysql/toolkit/script/__init__.py
@@ -1,6 +1,6 @@
 from mysql.toolkit.script.split import SplitCommands
 from mysql.toolkit.script.dump import dump_commands
-from mysql.toolkit.script.execute import SQLScript
+from mysql.toolkit.script.script import SQLScript
 
 
 __all__ = ['SplitCommands', 'dump_commands', 'SQLScript']
